refactor(api-gen): move css class constants to styling directory
Refactoring towards eliminating catch-all directories like "constants"
