+++ bazel/api-gen/rendering/styling/css-classes.ts
@@ -1,3 +1,11 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
 // TODO(jelbourn): all of these CSS classes should use the `docs-` prefix.
 
 export const PARAM_KEYWORD_CLASS_NAME = 'adev-param-keyword';
