+++ bazel/api-gen/rendering/templates/parameter.tsx
@@ -9,7 +9,7 @@
 import {h} from 'preact';
 import {ParameterEntryRenderable} from '../entities/renderables';
 import {RawHtml} from './raw-html';
-import {PARAM_GROUP_CLASS_NAME} from '../constants/html-classes';
+import {PARAM_GROUP_CLASS_NAME} from '../styling/css-classes';
 
 
 /** Component to render a function or method parameter reference doc fragment. */
